DelayedKalmanFilter

DelayedKalmanFilter.calculate_current_number_of_states

DelayedKalmanFilter.calculate_final_covariance_matrix

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point

DelayedKalmanFilter.calculate_steady_state_of_ode

DelayedKalmanFilter.distribution_prediction_at_given_time

DelayedKalmanFilter.get_current_and_past_mean

DelayedKalmanFilter.hes_ode!

DelayedKalmanFilter.hill_function

DelayedKalmanFilter.initialise_state_space_mean

DelayedKalmanFilter.initialise_state_space_variance

DelayedKalmanFilter.kalman_filter

DelayedKalmanFilter.kalman_filter_state_space_initialisation

DelayedKalmanFilter.kalman_prediction_step!

DelayedKalmanFilter.kalman_update_step!

DelayedKalmanFilter.predict_state_space_mean!

DelayedKalmanFilter.predict_state_space_variance!

DelayedKalmanFilter.time_constructor_function

DelayedKalmanFilter.update_stacked_state_space_mean

DelayedKalmanFilter.calculate_current_number_of_states — Method

Calculate the current number of states for a given observation time

DelayedKalmanFilter.calculate_final_covariance_matrix — Method

Return the covariance matrix at the current (last predicted) number of states

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point — Method

Calculates the negative log of the likelihood of our data given the paramters, using the Kalman filter. It uses the predictedobservationdistributions from the kalman_filter function. The entries of this array in the second and third columns represent the probability of the future observation of mRNA and Protein respectively, given our current knowledge.

Parameters

protein_at_observations::: Observed protein. The dimension is n x 2, where n is the number of observation time points. For each data set, the first column is the time, and the second column is the observed protein copy number at that time.

model_parameters::Vector{<:AbstractFloat}: An array containing the model parameters in the following order: P₀, h, μₘ, μₚ, αₘ, αₚ, τ.

measurementvariance : float. The variance in our measurement. This is given by Sigmae in Calderazzo et. al. (2018).

Returns

log_likelihood : float. The log of the likelihood of the data.

DelayedKalmanFilter.calculate_steady_state_of_ode — Method

Calculate the steady state of the Hes5 ODE system, for a specific set of parameters.

Arguments

P₀::Float64

h::Float64

μₘ::Float64

μₚ::Float64

αₘ::Float64

αₚ::Float64

Returns

steady_state_solution::Array{Float64,1}: A 2-element array, giving the steady state for the mRNA and protein respectively.

DelayedKalmanFilter.distribution_prediction_at_given_time — Method

Calculate the mean and variance of the Normal approximation of the state space for a given time point in the Kalman filtering algorithm and return it as a Normal distribution

DelayedKalmanFilter.get_current_and_past_mean — Method

Returns past time, current mean and past protein

DelayedKalmanFilter.hes_ode! — Method

Function which defines the HES5 ode system

DelayedKalmanFilter.hill_function — Method

Calculate the Hill function for a given protein molecule number, repression threshold, and hill coefficient.

Arguments

protein::Real

P₀::Real

h::Real

DelayedKalmanFilter.initialise_state_space_mean — Method

Initialse the state space mean for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.initialise_state_space_variance — Method

Initialse the state space variance for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.kalman_filter — Method

Perform a delay-adjusted non-linear stochastic Kalman filter based on observation of protein copy numbers. This implements the filter described by Calderazzo et al., Bioinformatics (2018).

Arguments

protein_at_observations::AbstractArray{<:Real}: Observed protein. The dimension is n x 2, where n is the number of observation time points. The first column is the time, and the second column is the observed protein copy number at that time. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters::ModelParameters: A ModelParameters object containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

measurement_variance::Real: The variance in our measurement. This is given by Sigma epsilon in Calderazzo et. al. (2018).

Returns

state_space_and_distributions::StateAndDistributions: TODO

DelayedKalmanFilter.kalman_filter_state_space_initialisation — Function

A function for initialisation of the state space mean and variance, and update for the "negative" times that are a result of the time delay. Initialises the negative times using the steady state of the deterministic system, and then updates them with kalmanupdatestep.

Arguments

protein_at_observations::AbstractArray{<:Real}: Observed protein. The dimension is n x 2, where n is the number of observation time points. The first column is the time, and the second column is the observed protein copy number at that time. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters::ModelParameters: A ModelParameters object containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

measurement_variance::Real: The variance in our measurement. This is given by Sigma epsilon in Calderazzo et. al. (2018).

Returns

state_and_distributions::StateAndDistributions: TODO

DelayedKalmanFilter.kalman_prediction_step! — Method

Perform the Kalman filter prediction about future observation, based on current knowledge i.e. current state space mean and variance. This gives rho{t+delta t-tau:t+delta t} and P{t+delta t-tau:t+delta t}, using the differential equations in supplementary section 4 of Calderazzo et al., Bioinformatics (2018), approximated using a forward Euler scheme.

Arguments

state_space::StateSpace: TODO

states::TimeConstructor: TODO

current_observation::Vector{<:AbstractFloat}: TODO

model_parameters::ModelParameters: A ModelParameters object containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

Returns

state_space::StateSpace: TODO

DelayedKalmanFilter.kalman_update_step! — Method

Perform the Kalman filter update step on the predicted mean and variance, given a new observation. This implements the equations at the beginning of page 4 in Calderazzo et al., Bioinformatics (2018). This assumes that the observations are collected at fixed time intervals.

Arguments

state_space::StateSpace: TODO

states::TimeConstructor: TODO

current_observation::AbstractArray{<:Real}: TODO

τ::Real: TODO

measurement_variance::Real: TODO

Returns

state_space::StateSpace: TODO

DelayedKalmanFilter.predict_state_space_mean! — Method

Predict state space mean to the next observation time index

DelayedKalmanFilter.predict_state_space_variance! — Method

Predict state space variance to the next observation time index

DelayedKalmanFilter.time_constructor_function — Function

Function which returns an instance of TimeConstructor, which is defines the numbers of states for various uses in the Kalman filter.

DelayedKalmanFilter.update_stacked_state_space_mean — Method

Take the stacked state space mean and return the updated state space mean in the original dimensions