Internals

Documentation for DelayedKalmanFilter.jl's internal functions.

See the Public Documentation section for documentation of the public interface.

Index

DelayedKalmanFilter.calculate_steady_state_of_ode

DelayedKalmanFilter.distribution_prediction

DelayedKalmanFilter.hes_ode!

DelayedKalmanFilter.hill_function

DelayedKalmanFilter.initialise_off_diagonals

DelayedKalmanFilter.initialise_state_space_mean

DelayedKalmanFilter.initialise_state_space_variance

DelayedKalmanFilter.kalman_filter_state_space_initialisation

DelayedKalmanFilter.kalman_prediction_step!

DelayedKalmanFilter.kalman_update_step!

DelayedKalmanFilter.predict_state_space_mean!

DelayedKalmanFilter.predict_variance_and_off_diagonals!

DelayedKalmanFilter.slow_get_mean_at_time

DelayedKalmanFilter.slow_get_variance_at_time

DelayedKalmanFilter.update_current_time_and_observation!

DelayedKalmanFilter.∂hill∂p

Internal Interface

DelayedKalmanFilter.calculate_steady_state_of_ode — Method

Calculate the steady state of the Hes5 ODE system, for a specific set of parameters.

Arguments

P₀::AbstractFloat

h::AbstractFloat

μₘ::AbstractFloat

μₚ::AbstractFloat

αₘ::AbstractFloat

αₚ::AbstractFloat

Returns

steady_state_solution::Array{AbstractFloat,1}: A 2-element array, giving the steady state for the mRNA and protein respectively.

DelayedKalmanFilter.distribution_prediction — Method

Calculate the mean and standard deviation of the Normal approximation of the state space for a given time point in the Kalman filtering algorithm and return it as an array.

DelayedKalmanFilter.hes_ode! — Method

Function which defines the HES5 ode system

DelayedKalmanFilter.hill_function — Method

Calculate the Hill function for a given protein molecule number, repression threshold, and hill coefficient.

Arguments

protein::AbstractFloat

P₀::AbstractFloat

h::AbstractFloat

DelayedKalmanFilter.initialise_off_diagonals — Method

Initialse the off diagonals for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.initialise_state_space_mean — Method

Initialise the state space mean for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.initialise_state_space_variance — Method

Initialse the state space variance for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.kalman_filter_state_space_initialisation — Method

A function for initialisation of the state space mean and variance, and update for the "negative" times that are a result of the time delay. Initialises the negative times using the steady state of the deterministic system, and then updates them with kalmanupdatestep.

Arguments

protein_at_observations: Observed protein. The dimension is n x 2, where n is the number of observation time points. The first column is the time, and the second column is the observed protein copy number at that time. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters: An array containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

observation_transform: A 1 x 2 matrix corresponding to the transformation from ob

served data to molecule number, for mRNA and protein respectively.

measurement_variance: The variance in our measurement. This is given by Sigma epsilon in Calderazzo et. al. (2018).

Returns

system_state::SystemState:

predicted_observation_distributions::Matrix{Float64}: A matrix of size N x 2. The entries in the first column are the predicted state

space mean and the entries in the second column are the predicted state space variance for each observation.

DelayedKalmanFilter.kalman_prediction_step! — Method

Perform the Kalman filter prediction about future observation, based on current knowledge i.e. current state space mean and variance. This gives rho{t+Δt-τ:t+Δt} and P{t+Δt-τ:t+Δt}, using the differential equations in supplementary section 4 of Calderazzo et al., Bioinformatics (2018), approximated using a forward Euler scheme.

Arguments

system_state::SystemState:

model_parameters: An array containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

Returns

system_state::SystemState

DelayedKalmanFilter.kalman_update_step! — Method

Perform the Kalman filter update step on the predicted mean and variance, given a new observation. This implements the equations at the beginning of page 4 in Calderazzo et al., Bioinformatics (2018). This assumes that the observations are collected at fixed time intervals.

Arguments

system_state::SystemState

measurement_variance: The variance which defines the measurement error, it is $Σ_ϵ$ in the equation $y = Fx + Σ_ϵ$.

observation_transform: A 1 x 2 matrix corresponding to the transformation from observed data to molecule number, for mRNA and protein

respectively.

Returns

system_state::SystemState

DelayedKalmanFilter.predict_state_space_mean! — Method

Predict state space mean to the next observation time index

DelayedKalmanFilter.predict_variance_and_off_diagonals! — Method

Predict state space variance to the next observation time index. There are three separate steps in the variance prediction: (1) integrate P(t,s) to P(t+nΔt,s) for s = t-τ:t, where nΔt is the number of hidden states. (propagateexistingoffdiagonals!()) (2) Integrate P(t,t) to P(t+nΔt,t+nΔt). (propagatevariance!()) (3) Integrate P(t,t) to P(t,t+nΔt) for t in t:t+nΔt-1. (propagatenewoff_diagonals!())

DelayedKalmanFilter.slow_get_mean_at_time — Method

This is used to get the mean at any time, for plotting purposes. Using the fast version will cause an extrapolation error for t < observationtimepoints[end] - τ

DelayedKalmanFilter.slow_get_variance_at_time — Method

This is used to get the variance at any time, for plotting purposes. Using the fast version will cause an extrapolation error for t < observationtimepoints[end] - τ

DelayedKalmanFilter.update_current_time_and_observation! — Method

Update the current time and observation of the system state

DelayedKalmanFilter.∂hill∂p — Method

The partial derivative of the Hill function with respect to the protein molecule number.

Arguments

protein::AbstractFloat

P₀::AbstractFloat

h::AbstractFloat