Internals

Documentation for DelayedKalmanFilter.jl's internal functions.

See the Public Documentation section for documentation of the public interface.

Index

DelayedKalmanFilter.calculate_current_number_of_states

DelayedKalmanFilter.calculate_final_covariance_matrix

DelayedKalmanFilter.calculate_steady_state_of_ode

DelayedKalmanFilter.distribution_prediction_at_given_time

DelayedKalmanFilter.hes_ode!

DelayedKalmanFilter.hill_function

DelayedKalmanFilter.initialise_state_space_mean

DelayedKalmanFilter.initialise_state_space_variance

DelayedKalmanFilter.kalman_filter_state_space_initialisation

DelayedKalmanFilter.kalman_prediction_step!

DelayedKalmanFilter.kalman_update_step!

DelayedKalmanFilter.predict_state_space_mean!

DelayedKalmanFilter.predict_state_space_variance!

DelayedKalmanFilter.state_space_mean_indexer

DelayedKalmanFilter.state_space_variance_indexer

DelayedKalmanFilter.time_constructor_function

Internal Interface

DelayedKalmanFilter.calculate_current_number_of_states — Method

Calculate the current number of states for a given observation time

DelayedKalmanFilter.calculate_final_covariance_matrix — Method

Return the covariance matrix at the current (last predicted) number of states

DelayedKalmanFilter.calculate_steady_state_of_ode — Method

Calculate the steady state of the Hes5 ODE system, for a specific set of parameters.

Arguments

P₀::Float64

h::Float64

μₘ::Float64

μₚ::Float64

αₘ::Float64

αₚ::Float64

Returns

steady_state_solution::Array{Float64,1}: A 2-element array, giving the steady state for the mRNA and protein respectively.

DelayedKalmanFilter.distribution_prediction_at_given_time — Method

Calculate the mean and variance of the Normal approximation of the state space for a given time point in the Kalman filtering algorithm and return it as a Normal distribution

DelayedKalmanFilter.hes_ode! — Method

Function which defines the HES5 ode system

DelayedKalmanFilter.hill_function — Method

Calculate the Hill function for a given protein molecule number, repression threshold, and hill coefficient.

Arguments

protein::Real

P₀::Real

h::Real

DelayedKalmanFilter.initialise_state_space_mean — Method

Initialse the state space mean for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.initialise_state_space_variance — Method

Initialse the state space variance for a given set of time states and the ODE system's steady state

DelayedKalmanFilter.kalman_filter_state_space_initialisation — Function

A function for initialisation of the state space mean and variance, and update for the "negative" times that are a result of the time delay. Initialises the negative times using the steady state of the deterministic system, and then updates them with kalmanupdatestep.

Arguments

protein_at_observations::AbstractArray{<:Real}: Observed protein. The dimension is n x 2, where n is the number of observation time points. The first column is the time, and the second column is the observed protein copy number at that time. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters::ModelParameters: A ModelParameters object containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

states::TimeConstructor: A TimeConstructor whose fields define various values relevant to number of states, e.g. discretedelay, totalnumberofstates, etc.

observation_transform: A 1 x 2 matrix corresponding to the transformation from observed data to molecule number, for mRNA and protein respectively.

measurement_variance::Real: The variance in our measurement. This is given by Sigma epsilon in Calderazzo et. al. (2018).

Returns

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively.

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay.

predicted_observation_distributions::Array{Normal{Float64}}: An array of length n, whose entries are Normal distributions with mean and variance equal to the state space mean and variance predictions for the corresponding time point.

DelayedKalmanFilter.kalman_prediction_step! — Method

Perform the Kalman filter prediction about future observation, based on current knowledge i.e. current state space mean and variance. This gives rho{t+delta t-tau:t+delta t} and P{t+delta t-tau:t+delta t}, using the differential equations in supplementary section 4 of Calderazzo et al., Bioinformatics (2018), approximated using a forward Euler scheme.

Arguments

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively.

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay.

states::TimeConstructor: A TimeConstructor whose fields define various values relevant to number of states, e.g. discretedelay, totalnumberofstates, etc.

current_observation::Vector{<:AbstractFloat}: The current time point and protein observation which acts as the initial condition for the prediction.

model_parameters::ModelParameters: A ModelParameters object containing the model parameters in the following order: repression threshold, hill coefficient, mRNA degradation rate,protein degradation rate, basal transcription rate, translation rate, time delay.

Returns

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively. With each prediction new values are saved to the relevant entries in the matrix.

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay. With each prediction new values are saved to the relevant entries in the matrix.

DelayedKalmanFilter.kalman_update_step! — Method

Perform the Kalman filter update step on the predicted mean and variance, given a new observation. This implements the equations at the beginning of page 4 in Calderazzo et al., Bioinformatics (2018). This assumes that the observations are collected at fixed time intervals.

Arguments

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively.

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay.

states::TimeConstructor: A TimeConstructor whose fields define various values relevant to number of states, e.g. discretedelay, totalnumberofstates, etc.

current_observation::Vector{<:AbstractFloat}: The current time point and protein observation which acts as the initial condition for the prediction.

τ::Real: The time delay parameter, model_parameters[7].

measurement_variance::Real: The variance which defines the measurement error, it is $Σ_ϵ$ in the equation $y = Fx + Σ_ϵ$.

Returns

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively. At each update step the relevant entries in the matrix are updated according to the update defined in Calderazzo et al., Bioinformatics (2018).

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay. At each update step the relevant entries in the matrix are updated according to the update defined in Calderazzo et al., Bioinformatics (2018).

DelayedKalmanFilter.predict_state_space_mean! — Method

Predict state space mean to the next observation time index

DelayedKalmanFilter.predict_state_space_variance! — Method

Predict state space variance to the next observation time index. There are three separate steps in the variance prediction: (1) integrate P(t,s) to P(t+nΔt,s) for s = t-τ:t, where nΔt is the number of hidden states. (horizontal/vertical) (2) Integrate P(t,t) to P(t+nΔt,t+nΔt). (diagonal) (3) Integrate P(t,t) to P(t,t+nΔt) for t in t:t+nΔt-1. (horizontal/vertical)

DelayedKalmanFilter.state_space_mean_indexer — Method

Helper function to return the state space mean for a given (real valued) time between t-τ and t+Δobs, where t is the current number of states.

DelayedKalmanFilter.state_space_variance_indexer — Method

Helper function to return the state space variance for two given (real valued) times t₁ and t₂, between t-τ and t+Δobs, where t is the current number of states.

DelayedKalmanFilter.time_constructor_function — Function

Function which returns an instance of TimeConstructor, which is defines the numbers of states for various uses in the Kalman filter.