Public Documentation

Documentation for DelayedKalmanFilter.jl's public interface.

See the Internals section for documentation of internal functions.

Index

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point

DelayedKalmanFilter.kalman_filter

Public Interface

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point — Function

Calculates the log likelihood of our data given the parameters, using a delay adjusted extended Kalman filter. It uses the predicted_observation_distributions from the kalman_filter function. The entries of this array in the second and third columns represent the probability of the future observation of mRNA and protein respectively, given our current knowledge.

Parameters

protein_at_observations::Matrix{<:AbstractFloat}: Observed protein. The dimension is N x 2, where N is the number of observation time points. For each data set, the first column is the time, and the second column is the observed protein copy number at that time.

model_parameters::Vector{<:AbstractFloat}: An array containing the model parameters in the following order: P₀, h, μₘ, μₚ, αₘ, αₚ, τ.

measurement_variance::Float: The variance in our measurement. This is given by $Σ_ϵ$ Sigma_e in Calderazzo et. al. (2018).

Returns

log_likelihood::AbstractFloat.

DelayedKalmanFilter.kalman_filter — Function

kalman_filter(
    protein_at_observations::Matrix{<:AbstractFloat},
    model_parameters::Vector{<:AbstractFloat},
    measurement_variance::AbstractFloat,
)

A Kalman filter for a delay-adjusted non-linear stochastic process, based on observation of protein copy numbers. This implements the filter described by Calderazzo et al., Bioinformatics (2018).

Arguments

protein_at_observations::Matrix{<:AbstractFloat}: Observed protein. The dimension is N x 2, where N is the number of observation time points. The first column is the time, and the second column is the observed protein copy number. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters::Vector{<:AbstractFloat}: A vector containing the model parameters in the following order: repression threshold (P₀), hill coefficient (h), mRNA degradation rate (μₘ), protein degradation rate (μₚ), basal transcription rate (αₘ), translation rate (αₚ), and time delay (τ).

measurement_variance::AbstractFloat: The variance in our measurement. This is given by $Σ_ϵ$ in Calderazzo et. al. (2018).

Returns

state_space_mean::Matrix{<:AbstractFloat}: An N x 3 matrix, where N is the total number of states. The columns are time, mRNA and protein respectively.

state_space_variance::Matrix{<:AbstractFloat}: An 2N x 2N matrix where N is the total number of states. It is constructed as a 2 x 2 block matrix, where the blocks give the covariance of (mRNA, mRNA), (mRNA, protein), (protein, mRNA), and (protein, protein) for all times (t,s) where abs(t -s) <= τ, the transcriptional time delay.

predicted_observation_distributions::Array{Normal{Float64}}: An array of length N, whose entries are Normal distributions with mean and variance equal to the state space mean and variance predictions for the corresponding time point.

Example

julia> using DelayedKalmanFilter

julia> protein = [0. 105.; 10. 100.; 20. 98.]; # times are 0., 10., 20., and protein levels are 105., 100., and 98. respectively

julia> model_parameters = [100.0, 5.0, 0.1, 0.1, 1.0, 1.0, 15.0];

julia> measurement_variance = 1000.0;

julia> ss_mean, ss_var, _, distributions = kalman_filter(
           protein,
           model_parameters,
           measurement_variance
       );

julia> distributions[1]
Distributions.Normal{Float64}(μ=77.80895986786031, σ=93.7064351407417)