Public Documentation

Documentation for DelayedKalmanFilter.jl's public interface.

See the Internals section for documentation of internal functions.

Index

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point

DelayedKalmanFilter.kalman_filter

Public Interface

DelayedKalmanFilter.calculate_log_likelihood_at_parameter_point — Function

Calculates the log likelihood of our data given the parameters, using a delay adjusted extended Kalman filter. It uses the predicted_observation_distributions from the kalman_filter function. The entries of this array in the second and third columns represent the probability of the future observation of mRNA and protein respectively, given our current knowledge.

Parameters

protein_at_observations::Matrix{<:AbstractFloat}: Observed protein. The dimension is N x 2, where N is the number of observation time points. For each data set, the first column is the time, and the second column is the observed protein copy number at that time.

model_parameters::Vector{<:AbstractFloat}: An array containing the model parameters in the following order: P₀, h, μₘ, μₚ, αₘ, αₚ, τ.

measurement_variance::Float: The variance in our measurement. This is given by $Σ_ϵ$ Sigma_e in Calderazzo et. al. (2018).

Returns

log_likelihood::AbstractFloat.

DelayedKalmanFilter.kalman_filter — Function

kalman_filter(
    protein_at_observations::Matrix{T},
    model_parameters::Vector{T},
    measurement_variance::T;
    off_diagonal_steps::Integer = 10,
) where T

A Kalman filter for a delay-adjusted non-linear stochastic process, based on observation of protein copy numbers. This implements the filter described by Calderazzo et al., Bioinformatics (2018).

Arguments

protein_at_observations: Observed protein. The dimension is N x 2, where N is the number of observation time points. The first column is the time, and the second column is the observed protein copy number. The filter assumes that observations are generated with a fixed, regular time interval.

model_parameters: A vector containing the model parameters in the following order: repression threshold (P₀), hill coefficient (h), mRNA degradation rate (μₘ), protein degradation rate (μₚ), basal transcription rate (αₘ), translation rate (αₚ), and time delay (τ).

measurement_variance: The variance in our measurement. This is given by $Σ_ε$ in Calderazzo et. al. (2018).

off_diagonal_timestep: The number of off-diagonal time steps. This parameter determines how many DDE's we need to solve in the predict_variance_and_off_diagonals! function.

Returns

system_state::SystemState:

predicted_observation_distributions::Matrix{<:AbstractFloat}: A matrix of size N x 2. The entries in the first column are the predicted state

space mean and the entries in the second column are the predicted state space variance for each observation.

Example

julia> using DelayedKalmanFilter

julia> protein = [0. 105.; 10. 100.; 20. 98.]; # times are 0., 10., 20., and protein levels are 105., 100., and 98. respectively

julia> model_parameters = [300.0, 4.0, 0.1, 0.1, 1.0, 1.0, 15.0];

julia> measurement_variance = 1000.0;

julia> system_state, distributions = kalman_filter(
         protein,
         model_parameters,
         measurement_variance
       );

julia> distributions[1, :]
2-element Vector{Float64}:
    98.83526916576103
 10883.526916576104